+++ include/zephyr/mgmt/hawkbit.h
@@ -89,62 +89,6 @@ int hawkbit_set_custom_data_cb(hawkbit_config_device_data_cb_handler_t cb);
 int hawkbit_init(void);
 
 /**
- * @brief Runs hawkBit probe and hawkBit update automatically
- *
- * @details The hawkbit_autohandler handles the whole process
- * in pre-determined time intervals.
- *
- * @param auto_reschedule If true, the handler will reschedule itself
- */
-void hawkbit_autohandler(bool auto_reschedule);
-
-/**
- * @brief Wait for the autohandler to finish.
- *
- * @param events Set of desired events on which to wait. Set to ::UINT32_MAX to wait for the
- *               autohandler to finish one run, or BIT() together with a value from
- *               ::hawkbit_response to wait for a specific event.
- * @param timeout Waiting period for the desired set of events or one of the
- *                special values ::K_NO_WAIT and ::K_FOREVER.
- *
- * @retval HAWKBIT_OK if success.
- * @retval HAWKBIT_NO_RESPONSE if matching events were not received within the specified time
- * @retval HAWKBIT_NETWORKING_ERROR fail to connect to the hawkBit server.
- * @retval HAWKBIT_UNCONFIRMED_IMAGE image is unconfirmed.
- * @retval HAWKBIT_PERMISSION_ERROR fail to get the permission to access the hawkBit server.
- * @retval HAWKBIT_METADATA_ERROR fail to parse or to encode the metadata.
- * @retval HAWKBIT_DOWNLOAD_ERROR fail while downloading the update package.
- * @retval HAWKBIT_UPDATE_INSTALLED if an update was installed. Reboot is required to apply it.
- * @retval HAWKBIT_NO_UPDATE if no update was available.
- * @retval HAWKBIT_CANCEL_UPDATE update was cancelled.
- * @retval HAWKBIT_NOT_INITIALIZED if hawkBit is not initialized.
- * @retval HAWKBIT_PROBE_IN_PROGRESS if probe is currently running.
- */
-enum hawkbit_response hawkbit_autohandler_wait(uint32_t events, k_timeout_t timeout);
-
-/**
- * @brief Cancel the run of the hawkBit autohandler.
- *
- * @return a value from k_work_cancel_delayable().
- */
-int hawkbit_autohandler_cancel(void);
-
-/**
- * @brief Set the delay for the next run of the autohandler.
- *
- * @details This function will only delay the next run of the autohandler. The delay will not
- * persist after the autohandler runs.
- *
- * @param timeout The delay to set.
- * @param if_bigger If true, the delay will be set only if the new delay is bigger than the current
- * one.
- *
- * @retval 0 if @a if_bigger was true and the current delay was bigger than the new one.
- * @retval otherwise, a value from k_work_reschedule().
- */
-int hawkbit_autohandler_set_delay(k_timeout_t timeout, bool if_bigger);
-
-/**
  * @brief The hawkBit probe verify if there is some update to be performed.
  *
  * @retval HAWKBIT_NETWORKING_ERROR fail to connect to the hawkBit server.
@@ -330,6 +274,70 @@ uint32_t hawkbit_get_poll_interval(void);
 int hawkbit_reset_action_id(void);
 
 /**
+ * @brief hawkBit autohandler.
+ * @defgroup hawkbit_autohandler hawkBit autohandler
+ * @ingroup hawkbit
+ * @{
+ */
+
+/**
+ * @brief Runs hawkBit probe and hawkBit update automatically
+ *
+ * @details The hawkbit_autohandler handles the whole process
+ * in pre-determined time intervals.
+ *
+ * @param auto_reschedule If true, the handler will reschedule itself
+ */
+void hawkbit_autohandler(bool auto_reschedule);
+
+/**
+ * @brief Wait for the autohandler to finish.
+ *
+ * @param events Set of desired events on which to wait. Set to ::UINT32_MAX to wait for the
+ *               autohandler to finish one run, or BIT() together with a value from
+ *               ::hawkbit_response to wait for a specific event.
+ * @param timeout Waiting period for the desired set of events or one of the
+ *                special values ::K_NO_WAIT and ::K_FOREVER.
+ *
+ * @retval HAWKBIT_OK if success.
+ * @retval HAWKBIT_NO_RESPONSE if matching events were not received within the specified time
+ * @retval HAWKBIT_NETWORKING_ERROR fail to connect to the hawkBit server.
+ * @retval HAWKBIT_UNCONFIRMED_IMAGE image is unconfirmed.
+ * @retval HAWKBIT_PERMISSION_ERROR fail to get the permission to access the hawkBit server.
+ * @retval HAWKBIT_METADATA_ERROR fail to parse or to encode the metadata.
+ * @retval HAWKBIT_DOWNLOAD_ERROR fail while downloading the update package.
+ * @retval HAWKBIT_UPDATE_INSTALLED if an update was installed. Reboot is required to apply it.
+ * @retval HAWKBIT_NO_UPDATE if no update was available.
+ * @retval HAWKBIT_CANCEL_UPDATE update was cancelled.
+ * @retval HAWKBIT_NOT_INITIALIZED if hawkBit is not initialized.
+ * @retval HAWKBIT_PROBE_IN_PROGRESS if probe is currently running.
+ */
+enum hawkbit_response hawkbit_autohandler_wait(uint32_t events, k_timeout_t timeout);
+
+/**
+ * @brief Cancel the run of the hawkBit autohandler.
+ *
+ * @return a value from k_work_cancel_delayable().
+ */
+int hawkbit_autohandler_cancel(void);
+
+/**
+ * @brief Set the delay for the next run of the autohandler.
+ *
+ * @details This function will only delay the next run of the autohandler. The delay will not
+ * persist after the autohandler runs.
+ *
+ * @param timeout The delay to set.
+ * @param if_bigger If true, the delay will be set only if the new delay is bigger than the current
+ * one.
+ *
+ * @retval 0 if @a if_bigger was true and the current delay was bigger than the new one.
+ * @retval otherwise, a value from k_work_reschedule().
+ */
+int hawkbit_autohandler_set_delay(k_timeout_t timeout, bool if_bigger);
+
+/**
+ * @}
  * @}
  */
